+++ src/shared/index.ts
@@ -104,6 +104,8 @@ export { useIsWindowLg } from "./hooks/useIsWindowLg";
 //Posts
 export { useSearchPost } from "./hooks/posts/useSearchPost";
 export { usePostForm } from "./hooks/posts/usePostForm";
+export { usePostInteraction } from "./hooks/posts/usePostInteraction";
+export { useCommentInteraction } from "./hooks/posts/comment/useCommentInteraction";
 
 //Apply
 export { useCreateApply } from "./hooks/posts/apply/useCreateApply";
